Modificado el sistema de combate, ahora devuelve un double en las criaturas
Añadido el atributo barrera, con sus getter y setter

Añadida funcionalidad al metodo de defensa, Mago y Guerrero

+++ src/modelo/criaturas/AccionesCriaturas.java
@@ -22,7 +22,7 @@ public interface AccionesCriaturas {
 	 *
 	 * @return the int
 	 */
-	public int atacar(Personaje objetivo);
+	public double atacar(Personaje objetivo);
 
 	/**
 	 * Defender.
